Fix RC4 return types, TLS return, RC4 crypt-key refinement use
When cryptogrpaphy was moved to an extension from the RXIARG-based API,
an error was made in translating the RC4 function from integer args to
name-based args.  In a place where "crypt-key" was intended (an ARG()),
"key" was used (a REF()).

This corrects that problem, as well as a couple of return-result
enablement issues.