Make HELP simulate soft quoting, but actually hard quote
It's nice if HELP allows one to use expressions, as HELP (...) telling
you that it's a literal GROUP! is pedantic and not very useful.  But
if this feature is accomplished with soft quoting, it runs into trouble
with an enfix operator which hard quotes its left...which outranks
an operation that soft quotes its right.  This causes problems e.g.
with HELP DEFAULT.

(Note that this is why it is usually suggested that enfix operators
soft quote their left, if they are a generic quoting operator.  DEFAULT
is special, and has to hard-quote.)

Hence this just goes ahead and makes HELP hard quote, but if it gets a
GET-WORD! or GET-PATH! or GROUP! it reduces it.  That gives the same
effect as soft quoting, but out-prioritizes any enfix-left-quoters.

Also changes the misleading argument name "word" to "topic", and
removes some superflous stale comments.