Add IDENTITY, have GROUP! not affect VALUE_FLAG_UNEVALUATED
This changes the logic of how the evaluated/unevaluated bit is done
such that groups don't affect it.  Hence `(1 2 3)` would come back
with 3 and consider it to be "unevaluated", while `(1 + 2)` would
give back a 3 and consider it evaluated.

This means that should IF not welcome evaluated integers and you need
to subvert that, you can write `if condition (semiquote do [1 + 1])`.
Otherwise you're forced to write `if condition semiquote do [1 + 1]`.
This would limit the caller's flexiblity of grouping for readability.

However, putting a value in a GROUP! was the previous way of forcing
the evaluated bit onto things.  Rather than introduce a SEMIEVAL
function, this observes that really what SEMIEVAL is is the Identity
Function:

https://en.wikipedia.org/wiki/Identity_function
https://stackoverflow.com/q/3136338

Rather than use a silly name for the identity function, this just goes
ahead and uses "identity".  Despite being nounish, it is standard,
and contexts that don't use it can use it as a variable name (as with
other things... HEAD, CASE, etc.)  Then /QUOTE is added as a refinement
with SEMIQUOTE just a specialization.

Though SEMIQUOTE is still a silly name in its own right, it is at
least not going to be taken for anything else.  So it is being kept
until there's a better suggestion.  (UNEVAL perhaps?)