Fix EITHER w/o branch in %host-start.r, conditional compliance
This brings conditional usages into compliance with the proposed rule
where literals may be used as branches in ordinary IF statements, but
not things that *evaluate* to literals.  The code to enforce this
(which was used to find these cases) is not yet included.

Fixes one problem found from not prohibiting this...where an EITHER
was missing its second branch.  Hence it spuriously picked up a
subsequent SET-WORD! statement.

See this discussion:

https://forum.rebol.info/t/212/