IS_FALSEY and IS_TRUTHY vs. IS_CONDITIONAL_TRUE
IS_FALSE and IS_TRUE from R3-Alpha were defined to match the TRUE?
and FALSE? tests, meaning "that which would lead an IF to take or not
take its branch":

https://github.com/rebol/rebol/blob/25033f897b2bd466068d7663563cd3ff64740b94/src/include/sys-value.h#L622

This ambiguity was recently excised by making TRUTHY? and FALSEY? due
to growing clout of these "goofy" terms in formal programming culture:

https://developer.mozilla.org/en-US/docs/Glossary/Truthy

But IS_CONDITIONAL_TRUE meaning "that which will make an IF take its
branch" has gotten more complex, by trying to be sensitive to the
common mistake of using a BLOCK! literal for a condition.  Hence,
although `[x = 1]` may be "truthy" in the sense that a *variable*
holding that value will be true, it's not "conditionally true"
because there's protection against using that block literally in an IF.
It is a known stumbling point for new users, and has proven useful
to catch the mistake even for not-new-ones (e.g. me).

Previously, this separation was indicated by IS_CONDITIONAL_TRUE vs.
IS_CONDITIONAL_TRUE_SAFE.  But now all "conditionals" potentially
use the safe form (based on refinements).  So it makes more sense to
break this up into TRUTHY, FALSEY, and an arity-2 IS_CONDITIONAL_TRUE,
which gets a boolean parameter of whether that mode of conditional is
enforcing safety.