Merge /ONLY and /OPT on conditionals, branch checks
The original meaning of the /ONLY refinement on conditionals was to
return blocks-as-blocks, e.g. `if/only true [print "hi"]` would be
the block `[print "hi"]`.  This feature never wound up being very
widely used, and could be implemented as `if true [[print "hi"]]`,
which is actually more succinct.

A more important form of "/only" arose (which carried the name "/opt",
meaning, "do not blankify the result of a taken branch which ends up
being void".

For a full discussion of why the former meaning of /ONLY is being
dropped, and the impliciations of /OPT taking it over, see:

https://forum.rebol.info/t/212/11

This also includes an early form of the "branch checks" which ensure
that an evaluated branch is always a block, while still permitting
non-evaluated literals to serve as branches.  It is a work in progress,
having the unfortunate property that the check cannot be inherited by
functions which wrap a conditional.

Since the rule is applied uniformly, note this means that CASE
statements like the following no longer work:

    case [
        true blank
    ]

One would have to write

    case [
        true [blank]
    ]

However, `case [true {foo}]` or `case [true 2]` are legal, as is
`two: [1 + 1] | case [true two]`.  Effectively it means the only
evaluations that are legal in branch slots are double evaluations
which will run code that's not visible in source...never a single
evaluation.  This gives more certainty to the developer (and the
reader of the code) about what is going on.