Limit Left-Hand Enfixed Normal Fulfillment to "one expression"
When an enfixed function has a "normal" (non-tight) first argument,
the initial implementation would hold lookahead in a suspended state,
leading to full completion of arguments on the left hand side.

This "full" completion led to undesirable effects, such as ELSE's
"Dark Corner".  For instance:

    print if false ["a"] else ["b"]

This would be interpeted as `(print if false ["a"]) else ["b"]`  This
is because when ELSE was first seen, it would be immediately after the
["a"] had been written into the output cell of IF's frame, creating a
situation like this with IF and PRINT on the stack above:

    print>if> *D_OUT* else ["b"]

Rather than take the D_OUT as its left hand side (the way a tight
argument would), it would yield...allowing the expression which had
requested one argument evaluation above to finish (IF), but with PRINT
still on the stack.  The next time ELSE would be seen would be its
next chance:

    print> *D_OUT* else ["b"]

But this situation was indistinguishable; there was no memory or bit
of the fact that ELSE had already had once chance to complete.  So it
would allow PRINT to complete.  This made non-tight left parameters a
bit like a backwards DO instead of a backwards DO/NEXT, which didn't
match the DO/NEXT nature of forward normal argument fulfillment.

The trick in this commit is to add a "dampening" bit onto the frame
above such enfixed functions, if that frame happens to be an argument
fulfillment frame.  Otherwise it falls back to "tight" behavior.

Among other things, this appears to resolve ELSE's dark corner, making
it a (hopefully) universally compatible replacement for EITHER,
SWITCH/DEFAULT, or ending CASEs with [... true ["some catch all"]]