Rename SET/LOOKBACK to SET/ENFIX, LOOKBACK? to ENFIXED?
This consolidates the two names for making functions that get their
first argument from the left under the single term "enfix", as it is
the more user-facing concept...while lookback is more of an internal
detail of the evaluator.

For full discussion, see:

https://forum.rebol.info/t/214/8

This commit also deletes some superfluous helpers like SET-INFIX and
SET-POSTFIX which were conveniences introduced prior to ENFIX's
existence (or before it could have existed, due to no left quoting
in enfixed functions).

Several comments are also brought up to date.