Hack to make HELP/DOC DEFAULT work
Ren-C refactored the Do_Core loop such that lookahead checking, used
to implement infix, could avoid "double-lookup".  e.g. if a lookahead
fetched a word's variable to see if it was an infix function, that
fetch would be reused as the DO of a frame was continued.

This fetching could theoretically be applied to paths, in order to
allow for refinements on infix functions.  But since R3-Alpha permits
GROUP!s to be in paths as part of the requirement for finding what
they dispatch to, evaluation of these groups can have side-effects.
And one does not want to have side-effects during a lookahead

(A DO/NEXT might not intend to run those groups ever, much less in
an out of order fashion.)

Due to this and other concerns, paths have not been able to properly
participate in enfixing at this time.  One problem with this, though,
is when a prefix function with refinements that hard quotes its first
argument wants to override an enfixed function on its right that wants
to left quote its first argument.  The usual ability of a hard quoting
left function to override the right is missed, because the path is
not known to be a quoting function...or a function at all.

This meant `HELP DEFAULT` could allow HELP to override DEFAULT's left
quoting behavior... but `HELP/DOC DEFAULT` obscured the hard-quoting
HELP...and DEFAULT "won" quoting left.  While this general problem is
going to take some redesign and rules to solve, a quick fix here just
peeks at the head of a path...fetches it to see if it's a function
that hard quotes its first argument, and treats it as it would treat
the word.

Note this can be broken as easily as `o: make object! [h: :help]` and
then o/h/doc, since the head will not be seen as a function.  It is
not a long term solution, it's just to assist with HELP/DOC.