EITHER-TEST and native specializations to power THEN/ELSE
This commit introduces a new conditional primitive, tentatively named
EITHER-TEST.  It is a ternary operator that takes a test (which can be
a datatype, typeset, function, block of words for types, etc.), a
value, and a branch.  If the value passes the test, the value is
returned...otherwise it runs the branch.

    >> either-test :even? 4 [print "it wasn't even" | 20]
    == 4

    >> either-test :even? 3 [print "it wasn't even" | 20]
    it wasn't even
    == 20

This function is then specialized and chained to produce the MAYBE,
MAYBE?, and ENSURE operations.

    >> maybe [integer! string!] "yes"
    == "yes"

    >> maybe [integer! string!] <no>
    == _

    >> ensure [integer! string!] "yes"
    == "yes"

    >> ensure [integer! string!] <no>
    ** Error indicating the callsite of ENSURE

Even more interesting is that when specialized with the test as :VOID?
this becomes the prefix version of THEN, and when specialized with the
test as :ANY-VALUE? (e.g. not void) it becomes the prefix version of
ELSE.  So all they have to do is be enfixed.

    then: enfix specialize 'either-test [
        test: :void?
    ]

    else: enfix specialize 'either-test [
        test: :any-value?
    ]

But specialization, while not particularly "heavyweight", is not
free.  So since THEN and ELSE are called so often (or hopefully will
be), this goes ahead and provides hand-coded native optimizations of
EITHER-TEST-VOID and EITHER-TEST-VALUE.

    then: enfix :either-test-void
    else: enfix :either-test-value