Enhance increment (++) and decrement (--) operators
The ++ and -- operators were not used very frequenty in Rebol, due to
their "non-Rebol" appearance.  They were prefix, would modify their
argument (quoted as a WORD!), and would return the prior result.

    >> x: 10

    >> -- x
    == 10

    >> x
    == 9

To a casual observer, it was not clear that x was being modified...
especially because it wasn't even being passed as a WORD!.  A proposal
favored by @GrahamChiu and @earl was that it might better be an
arity-2 infix operator which would quote its left hand side, and
allow incrementing or decrementing by any amount.  (Hence more like
C's += and -=)

    >> x: 10
    == 10

    >> x: ++ 3 * 3
    == 19

    >> x: -- 2
    == 17

    >> t: 12:00
    == 12:00

    >> t: ++ 3:00:42
    == 15:00:42

    >> s: [a b c d]
    == [a b c d]

    >> s: ++ 2
    == [c d]

If one wished for simple incrementing or decrementing, it doesn't
seem so hard to say `my-variable-name: ++ 1`, as the real goal is
to avoid repeating the variable name.

While the proposal was liked, the technical ability to implement it
wasn't available until relatively recently.  This adds it.  Though it
breaks previous behavior, it's likely not typical that prior usages
of ++ or -- were preceded by a SET-WORD! or SET-PATH!, so they should
hopefully be easy to catch.