Remove series skipping behavior from ++ and --
While R3-Alpha's ++ and -- were willing to do NEXT and BACK on series,
this is not a generic interaction of the + or - operators with series,
just as + is not endorsed in the box as a string concatenation
operator.

This commit makes the operators more generic in terms of "doing
whatever it is that + or - would do".

Fixes up some comments on conditionals, also.