Correct block matching block edge case, tidy Find_In_Array
All R3-Alpha arrays (theoretically) were terminated by a full
cell-sized REBVAL of type END!.  Users were never supposed to see these
"values".  But it was somewhat common to see them leak into userspace,
because the system was not especially noisy about their appearance.

Some algorithms depended on this implicit terminator, e.g. with code
like:

    while (VAL_TYPE(value) == REB_INTEGER)
       ++value;

This code would increment through a value array as long as it was an
integer, but be implicitly stopped before crossing the end because
VAL_TYPE() on the terminal cell would be REB_END, and not match.

Ren-C excised several dangers of this, by not having a "value" of type
END! to leak.  VAL_TYPE(value) would fail on terminator cells, and the
condition of being a terminator was reduced down to a single bit.  This
also permitted more compact forms of arrays, such as "singulars" which
can fit a single element array in a single series node...since it need
not carry the overhead of a full cell for a terminator.

Over time most every bit of code that depended on the old behavior was
fixed up, as it triggers a very clear error.  Though it has been a long
time since one of these categories of errors has been seen, there was
a latent case in Find_In_Array:

    >> find [this that] [that this]
    VAL_TYPE() called on END marker

While the enumeration of the item to search for was guarded by
NOT_END(), the parallel walk in the search target had no protection.
This adds that protection, and also goes ahead and tidies up the
Find_In_Array routine a bit to modern coding style.