Make ? and ?? conditional operators that don't evaluate blocks
Originally in Rebol2 and R3-Alpha, ? was a synonym for HELP.  This
seems wasteful for the language as a whole, when it's easy enough to
type HELP, or add it to the console-specific abbreviations as H (as
with Q for QUIT).

This is an experiment which makes ? an infix operator, that takes
a condition on the left and a value on the right--like an IF that won't
run blocks or functions.  As a complement, ?? is then taken as a
parallel to ELSE, which will also not run blocks or functions.
Both functions tolerate arbitrary evaluations, with none of the
safety features of conditionals that do execute blocks.

    >> 1 > 2 ? [hi] ?? [bye]
    == [bye]

    >> 1 < 2 ? [hi] ?? [bye]
    == [hi]

    >> 1 < 2 ? 3 + 3 ?? 4 + 4
    == 6

    >> 1 > 2 ? 3 + 3 ?? 4 + 4
    == 8

    >> 1 > 2 ? [a] ?? () ?? () ?? [x] ?? () ?? [y]
    == [x]