Add new CHOOSE native as replacement for old CASE/ONLY behavior
(suggestion by @draegtun)

One of the reasons for the enhanced protection on branches is the
somewhat unpredictable nature of how blocks are handled by control
constructs.  e.g.

    ssd: "some string data
    sbd: [some block data]

    value: case [
       (1 == 2) ssd
       (2 == 2) sbd
    ]

This would give rise to an error, because it would try to run `sbd`
as code.  Being able to abstract branches in this way is important
when customizing control constructs, but dangerous when choosing data
values when "double-evaluation" wasn't intended.

For this reason, CASE/ONLY used to mean "return the value 'as-is'".
This pattern was removed, but since it is useful this factors out the
code behind CASE into a routine shared with a new native called CHOOSE,
which for the above code would return the block `[some block data]`.

The /ALL refinement is kept, which actually means something more like
"/LAST" because you're making the last choice.  However, since this
shortcuts evaluations of conditions and values, it's probably best
to help people remember that's what's happening...that any side
effects from running all [] the expressions will happen.