Arity-1 Function Branches receive conditional arguments
This changes Run_Branch_Throws() to also take the condition which
triggered the branch.  This way, if the branch is an arity-1 function,
that condition can be provided as a parameter to the function:

    >> if false func [x] [print x | 304]
    ;-- no result

    >> if 1020 func [x] [print x | 304]
    1020
    == 304

It works with loops like WHILE as well, should the body be an arity-1
function, as well as all the conditionals like CASE, THEN, ELSE...
SWITCH is a little odd because there's no such thing as a function
literal, but if you compose a function into the cases block it will
currently do the same thing.

    >> switch 1 compose [1 (func [x] [print x])]
    1
    == _

Also, with "blankification" more uniform as a process, this moves it
into the handler for Run_Branch_Throws() to reduce repetition.  As a
result, the REB_R codes R_OUT_VOID_IF_UNWRITTEN_BLANK_IF_VOID and
R_OUT_BLANK_IF_VOID could be removed.

Effective applications of this feature aren't really explored yet.  But
with this change, there's enough information to give a not-completely
useless message when ensure fails, since it can at least tell you the
type of the failing variable.  So the /ERROR-HACK is removed, and it
demonstrates at least one practical use.