Repurpose UPDATE as the "anti-DEFAULT", rename old UPDATE
This defines UPDATE as being like DEFAULT.  But rather than rejecting
an assignment based on whether the current value of a variable is
blank/void, UPDATE rejects the assignment if the NEW value is blank or
void...keeping the old value in these cases.

    >> x: 10

    >> x: update ()
    >> print x
    10

    >> x: update [1 + 1]
    >> print x
    2

    >> x: update 20
    >> print x
    20

    >> x: update if 1 > 2 [1000]
    >> print x
    20

Because the old meaning of UPDATE looked to be an internal-use-only
port action called exclusively by WAKE-UP, this renames that action
to ON-WAKE-UP pending further understanding of its applications to
userspace code (not to mention further redesign of the port model).

Includes some streamlining of functions in %base-funcs.r, including
making DOES a specialization of FUNC as well as using specializations
of ENSURE-TEST to speed up some chained functions vs. using userspace
functions.