SET+GET drop /ANY+/OPT to use /ONLY, SET/ONLY => SET/SINGLE
SET/ANY was a poor name for what it did, but a suitable name was
difficult to find.  Ren-C started out trying /OPT, but it was not a
slam dunk.  Hence the legacy support for /ANY was kept turned on rather
than force a migration, so that /OPT could be allowed to settle, and
it wouldn't take two churns if a better name was found.

During that time, SET and GET evolved...notably that the non-OPT/ANY
form of GET would return a BLANK! in the case of an unset variable
rather than being an error.  This simplified several scenarios and
helped alleviate some pain caused by increased void-ing in the system.
A complementary operation GET-VALUE was added to trigger an error if
a variable was void, implemented as a CHAIN on top of GET/ONLY which
complained if the piped out result was void.

With a new framework of thinking about this, the /ANY or /OPT version
came to seem more like the "fundamental" version, upon which higher
level GETs and SETs were implemented.  This is what /ONLY has come
to mean, and which is being abbreviated as `*` in various places.

So you might think that a GET/ONLY is a GET, only without the
"blankification" applied to voids.  And SET/ONLY is a SET, only without
throwing in an extra parameter check that the value you are trying to
set to isn't void (which actually *unsets* the variable).

One hitch in this plan was that SET/ONLY has a meaning, introduced by
Red, as follows:

    >> set/only [a b] [c d]
    >> a
    == [c d]
    >> b
    == [c d]

So the operation is if the target is a block, and the source is a
block, you do not intend to set it itemwise (e.g. such that a is c
and b is d).  Rather, you meant to set all the targets in the block
to the same block value.

This option was not in R3-Alpha, and not used in the Ren-C codebase.
To free up /ONLY for the more useful definition, this renames that
functionality to /SINGLE.  (Name suggestion by @ingo, there are no
known uses of it so still changeable if someone has a better idea.)

Also speeds up GET-VALUE by modifying the chain to feed into a
native specialization.