Standardize molding code to common naming and formatting
Unlike other per-type dispatched operations, molding is not done with
a table of functions...rather it was done using a convoluted switch
statement.  The functions called and their parameters were rather
irregular.  This commit attempts to organize it in a more regular way
to make it easier to turn molding/forming into a table-driven process.

The "molded" flag was confusing, because it made it seem like the
operation was dealing with something that had already been molded.
But really it was a flag distinguishing Mold_Value from being invoked
for molding or forming.  Among the several clarifying changes, an
operation that could either Mold or Form is renamed Mold_Or_Form_XXX,
and the flag is inverted and renamed to "form".  This is then turned
into Mold_Value and Form_Value, instead of calling Mold_Value with a
mysterious REBOOL.

The molding structure is now always called `mo`, and there is a macro
for declaring it and clearing it, with all accesses through pointers.
The flags are switched over to being MOLD_FLAG_XXX with get/set/clear
operations, instead of "MOPT" and "GET_MOPT".

Beyond clarifying changes, this just does general modernization of the
code doing forming and molding, which has barely been touched in the
Ren-C effort.