RIGOROUS updates for MSVC build
This disables a few warnings, corrects a C++-style-no-arguments
function (e.g. foo()) to C-style-no-arguments function (e.g. foo(void))
and blanks out the "<TreatWarningAsError>" field in the project file
so it doesn't take the default of "/WX-" which would then get
overridden by the "/WX" of RIGOROUS later on the command line.