Move APPLY's safety check from native to %mezz-legacy.r
There was a check of APPLY's block parameter to see if it started with
a SET-WORD! or not, as a rough heuristic of whether someone is likely
using the R3-Alpha APPLY convention, or the new Ren-C convention.

The check was done in native code, but as it's really a legacy feature,
it's better to do this in %mezz-legacy.r by turning APPLY into an
adaptation of itself which throws in the check.  This makes it easier
to understand and/or to get rid of.

Further work on APPLY is needed to allow it to reference parameters
positionally.  (Some ideas of using SET-INTEGER! for this, e.g.
`1: foo` instead of hard-coding names of parameters, have been put
on the table.)