Use LodePNG for decoding as well as encoding, kill %u-png.c
Due to a bug in the R3-Alpha custom PNG code (%u-png.c) related to
encoding, Saphirion introduced a dependency on the lightweight single
file PNG encoder/decoder "LodePNG".  Including LodePNG meant including
a second implementation of inflate() and deflate(), which were already
present in Rebol as "zlib".  (%u-png.c reused this code)

So there were two PNG encoders and two zlibs in the EXE :-(

This commit brings LodePNG up to date with its master on GitHub, and
uses it for PNG decoding as well...getting rid of %u-png.c.  It also
uses the LodePNG option to plug in a custom zlib and not build its own.
This gets rid of the redundancy.

An optimization in the code makes it so that LodePNG uses a custom
allocator that replaces malloc(), yet puts the data in a series.  The
trick is that the series stores a pointer to itself at the head of the
data, which can be hacked off by means of increasing the "bias"
(unused capacity at the head of a series).  This permits buffers
allocated by LodePNG to be tweaked for use to hand back to users,
so no need to copy from malloc'd data into series data.

Because this "Rebserize()" feature is generically useful (and requires
low-level access to the memory structures) it is built as a generic
service in %m-pools.c.  Technically this commit does not really take
advantage of it yet, because IMAGE! does not use byte size series
at the moment.  It probably should--and it may be easier than changing
the Rebserize() and associated allocators to manage the various
issues involved in rounding the data and sizes.