Methodize MOLD so that STRUCT!'s mold code is in the extension
This brings the FFI extension commit closer to completion, by making
it possible to move the implementation of molding for types that have
been given `enum Reb_Kind` numbers, cell structures, and GC behavior
only in the core.

Moves many molding/forming behaviors (MF_Xxx) to live in the file with
their associated datatypes, instead of being in the %s-mold.c file.
It does not intentionally change mold behavior of any existing types.
So if behavior changes are observed, they are likely bugs.