Make DATE! without times and zones distinct from those with
R3-Alpha considered a date without a time, e.g. 12-Dec-2012, to be
semantically equivalent to one with a time at midnight.  It also
considered a date with a time zone of 0:00:00 to be the same as a
date without a time zone.

This changes dates to use two header bits to track whether it has a
zone or a time, and makes the debug build accessors stop access to
the VAL_NANO() (for time in nanoseconds) or the VAL_ZONE() when the
flags are not set.

There are several issues with the semantics of dates, and problems in
the implementation.  This commit does not address those...and due to
priorities, issues will be addressed on an as-needed basis.  The
general guidelines, however, are that there should be errors instead
of normalization when fields of a date are assigned that would make it
invalid...and that time-based math cannot be done on dates without
time components.