Specialization uses facades to suppress hidden keys in frames
The concept of a FRAME! is to enable the representation of a state of
a Rebol function invocation (or a potential invocation) of a Rebol
function, or specific invocation, as a context that resembles OBJECT!.
Not only is this useful in debugging, but in the design of custom
APPLY-like operations.

That model is complex to simulate for the user, because the operations
that do function compositions (SPECIALIZE, CHAIN, ADAPT, HIJACK...)
don't actually create distinct stack frames to run their code.  This
gives them a significant performance advantage over methods which would
rewrite the bodies of functions or otherwise execute more retriggering
levels.  An adaptation of a specialization of an adaptation of a
specialization would find each adaptation layer "perceiving" a
different number of parameters to the function.

In the initial implementation, no attempt was made to filter the number
of parameters seen in frame creation.  Hence if a FRAME! was made for
a function with 10 parameters that had 9 specialized out, it would
still appear to be a frame with 10 keys and values instead of 1.

Eventually, a ->phase field that is the currently represented FUNCTION!
for a given capture of a FRAME! REBVAL was introduced.  This commit
introduces a new concept of making specializations create a "facade"
pseudo-paramlist which hides the appropriate fields, and adapts the
CTX_KEYLIST() function to use that facade instead of either the
function's paramlist or the underlying function's paramlist--neither
of which would be appropriate.

Redundancy in Make_Frame_For_Function() & Specialize_Function_Throws()
was also noticed and removed.