USE=>native: performance + decouple from "closure" semantics
The R3-Alpha implementation of USE would build a closure (indefinite
extent lifetime for its arguments, refinements, and locals) based on
the given spec.  Then it would DO that function.

There's an inefficiency in this, because it constructs a function, that
takes time and also makes entities that need to be GC'd, when that
function is to be called only once.

This reduces the overhead of USE significantly by making it a native
and reuses the same code that copies and rebinds bodies of loops.
This helps USE not be specifically tied to the behavior or decisions
regarding indefinite or definite lifetimes in functions.  It also
puts the decision closer to behaviors which would be able to take
advantage of working virtual binding.