Make REMOVE-EACH respect locked/running/protected series bit
The various bits that mark a series read only have to be manually
checked by the C code that does mutating operations.  This manual
check was missing from the implementation of REMOVE-EACH in R3-Alpha.
This adds that check, which is only applicable in the shared loop
code for ANY-SERIES!, as MAP! and OBJECT! are not supported by the
REMOVE-EACH native.

Includes some superfluous code tidying.