Initialize facade for FFI routines prior to Make_Function()
The keylist used for a FRAME! depends on the phase that frame is in.
At an earlier moment or "outer layer" of specialization, it may use
a keylist that has some keys hidden...e.g. the "facade" for that
specialization.  All the fields must be present to be ultimately used
by the frame (to call the native, etc.) but enumeration via FOR-EACH
or other user mode constructs should not see them.

The optimized generator for FUNC and PROC creates meta-information that
actually uses frames, this generator is called Make_Paramlist_May_Fail.
Because the result of this routine is passed to Make_Function, it must
create the facade before creating such frames.  However, Make_Function
also allows callers to provide a facade of their own.

To avoid setting the facade twice (and possibly inconsistently) the
new invariant is that if Make_Function() receives a NULL facade, then
it must already be set in the paramlist, and it must be equal to the
paramlist itself.