Fix backwards assert, call to FLD_DIMENSION() on non-array
An assertion improperly tested that structure fields must have no name,
vs the correct assert that non-structure fields cannot have no name
(as they cannot be schemas).

Also, assertions were tightened so that FLD_DIMENSION() could not be
called on arrays, so this adds a check that a field is an array before
using it when embedding one structure's data into another.