REMOVE-EACH locks series until end, performs all removes atomically
The REMOVE-EACH primitive was unusual because it would perform mutation
on the very series it was iterating, based on the results of calling
into arbitrary user code.  This gives rise to a category of problems,
one very simple one being:

    data: copy [1 2 3 4]
    remove-each item data [
        protect data
        true
    ]

Since user code can transition a series from being mutable to being
immutable, this meant there'd have to be a check of whether that had
changed on every step of the REMOVE-EACH.  This isn't the only problem,
but it demonstrates the slippery slope one starts to be on if the
attempted answer is to add a "did-the-protection-status-change" test
on every loop iteration.

The need to constantly keep the input array in sync with the removals
also presented a performance problem, because each removal in a Rebol
series requires sliding up the remaining data to "close the gap".
Hence a series of length N that requested a removal on each step would
wind up doing N-1 "slides" of however much data was remaining, which
is effectively O(N^2) operations.

This motivated a rethinking of the native, and SERIES_INFO_RUNNING
(which was used to put temporary holds on blocks for frames) is now
SERIES_INFO_HOLD, so that it can more generally apply to enumerations
on strings or binaries.  Besides just being read-only during the life
of the REMOVE-EACH,, the implementation is changed to not reflect
any of the requested removals until the operation is completed--either
normally, by a BREAK, by a THROW, or via a fail().  All exit paths
then atomically perform the removals.

Different methods are used to accomplish this for arrays vs. strings or
binaries.  Since values in arrays have header bits which can be used,
the NODE_FLAG_MARKED (which is not used by GC for in user arrays) is
used to mark all the removals.  Move_Value and Derelativize then do
not propagate this flag, so it won't escape the use of the operation.
No such bit is available for strings or binaries, so they build new
underlying data arrays which are swapped in to back the input series,
with the original underlying data freed.