BREAK/WITH and CONTINUE/WITH in REMOVE-EACH, other fixes
The concept of BREAK/WITH is to finish the loop as if the body returned
a value before ceasing iteration.  In the case of REMOVE-EACH this
allows for a decision about the current iterated element's removal,
e.g. `BREAK/WITH TRUE` would remove it.  (The default BREAK effectively
acts as `BREAK/WITH ()`, which is as if the body opted out, hence there
will be no removal.)

CONTINUE/WITH is similar, allowing one to skip the rest of the current
iteration body but still be able to request removal of the element(s).

This adds proper support for the /WITH versions, and also fixes some
edge cases about how fail() would be handled...by deferring the
PUSH_TRAP until right before the iteration, when the finalization
routine can truly be safely called.