Restore tracking information for void, BAR!, BLANK!, LOGIC!
The debug build uses the payload of value cells that don't need a
payload to save the file and line number where the cell is initialized.
It then uses the "extra" field to store the DO_COUNT for the last time
that cell was written.

When the VAL_RESET_HEADER_EXTRA() optimization was added, this got lost
for those types.  This introduces VAL_RESET() which is a stronger
version of that function that assumes you want the track payload
initialized in the debug build...and restores the functionality.