Make core DO and APPLY actions hookable, implement TRACE as hook
Ren-C reimagined the evaluator in terms of acting on a C structure that
reflects the run state (a "REBFRM"), which could then be tied to a
Rebol value that introspected it (a "FRAME!").  This meant that the
entire evaluator was--effectively--a void function taking a single
parameter.  This function is called Do_Core().

One consequence of this is that the evaluator can be swapped out...
most likely for another evaluator that augments Do_Core() but uses it
as part of the implementation.  These augmentations could add tracing,
statistics, or debugging features.

Historically TRACE was implemented by setting Trace_Flags, and then the
choice to output trace information was made by doing `if (Trace_Flags)`
multiple times in the Do_Core() routine.  This commit institutes a new
strategy: to make all evaluator calls go through a level of dereference
and swap in `Do_Core_Traced()` as an augmented evaluator.

A second swap point is instituted to specifically hook the evaluation
of a function with its frame completed.  Due to the evaluator design,
this hook presently gets access to each phase of a composite function,
and has to simulate the processing of the R_XXX codes.  It may be that
this handling can be moved into Apply_Core, but the workings of some
`goto` instructions make that not feasible at this time.

While adding the level of dereference does have some performance
implications, it breaks relatively even due to the removal of branching
code based on flags.  Hence it is a generalized approach that allows
more stacks of behavior to be plugged in, and centralizes the code
that implements features like trace.